% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{CalculateGeneCV}
\alias{CalculateGeneCV}
\title{Calculate each gene's CV and write it into RNA assay}
\usage{
CalculateGeneCV(object, min.cell = 3, min.rvalue = -Inf, CNV_correction = F)
}
\arguments{
\item{object}{A Seurat object}

\item{CNV_correction}{Do CNV correction or not(For Cells which have CNV)}
}
\value{
Returns Seurat object with CV value in each gene in RNA assay
}
\description{
Calculate each gene's CV and write it into RNA assay
}
