% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ClusterCNV}
\alias{ClusterCNV}
\title{Clustering CNV to clusters}
\usage{
ClusterCNV(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns cluster result or Seurat object
}
\description{
Use UMAP or PCA to cluster CNV segments with weight to cluster cells
}
