% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{CNVRegionLocated}
\alias{CNVRegionLocated}
\title{Get CNV region located chromosome}
\usage{
CNVRegionLocated(object, region)
}
\arguments{
\item{object}{Some CNR object and Seurat object}

\item{region}{A character contain CNV region serial number}
}
\value{
Returns a character contain chromosomes which CNV regions located
}
\description{
Get CNV region located chromosome
}
