% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{FilterCNR}
\alias{FilterCNR}
\title{Filter CNR object}
\usage{
FilterCNR(CNR_obj, threshold = 2.5, topn = 10)
}
\arguments{
\item{CNR_obj}{A CNR object}

\item{threshold}{threshold of low quality cells}

\item{topn}{topn sites' depth to compare with median depth}
}
\value{
Returns a CNR object after filter
}
\description{
Remove some low quality cell in CNR object which major reads covered in a tiny region
}
