% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{GeneCNVRegion}
\alias{GeneCNVRegion}
\title{Write CNV region into RNA assay's each gene}
\usage{
GeneCNVRegion(object, gene_list = Human_hg38.bed, verbose = T)
}
\arguments{
\item{object}{A Seurat object}
}
\value{
Returns Seurat object with CNV region in each gene in RNA assay
}
\description{
Write CNV region into RNA assay's each gene
}
