% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{getCNVslot}
\alias{getCNVslot}
\title{get CNV assay slot}
\usage{
getCNVslot(object, slot)
}
\arguments{
\item{object}{A CNR or Seurat object}

\item{slot}{CNR, CNV, ploidyCNV or meta slot}
}
\value{
Returns a matrix or data.frame of slot
}
\description{
get CNR, CNV, ploidyCNV and meta slot from Seurat or CNR object
}
