% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{InferCNR}
\alias{InferCNR}
\title{Infer CNR from expression matrix}
\usage{
InferCNR(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns Seurat object or a list with InferCNR matrix and its metadata
}
\description{
Run InferCNR to expression matrix
}
