% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{mapCNR2Seurat}
\alias{mapCNR2Seurat}
\title{Map DNA CNR to Seurat InferCNV slot}
\usage{
mapCNR2Seurat(CNR, Seurat_obj, CNV.ratio = NULL)
}
\arguments{
\item{CNR}{A CNR object}

\item{Seurat_obj}{A Seurat object with InferCNV slot to be mapped}

\item{CNV.ratio}{Specify a CNV ratio if you know it}
}
\value{
Returns a CNR object with mapped slot
}
\description{
Map DNA CNR region to RNA to jointy analysis
}
