% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{MergeCNV}
\alias{MergeCNV}
\title{Merge CNR object and Seurat object's CNV}
\usage{
MergeCNV(
  ...,
  CNVlist = NULL,
  min.interval = 1e+05,
  scale = T,
  chr.use = paste0("chr", c(seq(1, 22), "X"))
)
}
\arguments{
\item{...}{Some CNR object and Seurat object}

\item{CNVlist}{A list contain CNR object and Seurat object}

\item{chr.use}{Used chromosome}
}
\value{
Returns a list with megered CNV and its metadata
}
\description{
Merge CNR object and Seurat object's CNV
}
