% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{NormalizeCNR}
\alias{NormalizeCNR}
\title{Normalize CNR to CNS}
\usage{
NormalizeCNR(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns CNS matrix or CNR object or Seurat object
}
\description{
Segment CNR matrix to segment matrix
}
