% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{parsimony_tree}
\alias{parsimony_tree}
\title{A estimated parsimony tree by CNV}
\usage{
parsimony_tree(segment)
}
\arguments{
\item{segment}{A segment}
}
\value{
Returns a estimated parsimony tree
}
\description{
Estimate a parsimony tree from CNV data
}
