% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{PloidyCNV}
\alias{PloidyCNV}
\title{Ploidy normalized CNR data}
\usage{
PloidyCNV(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns CNV matrix to CNR object or Seurat object
}
\description{
Ploidy normalized CNR matrix to CNV matrix
}
