% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{PlotCNR}
\alias{PlotCNR}
\title{Plot CNR}
\usage{
PlotCNR(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns pheatmap object or NA if specify file to draw
}
\description{
Plot DNA CNR or RNA CNR
}
