% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{PlotCNV}
\alias{PlotCNV}
\title{Plot CNV}
\usage{
PlotCNV(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns ggplot2 object or NULL if specify file to draw
}
\description{
Plot DNA CNV or RNA CNV
}
