% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotCNVRegion}
\alias{PlotCNVRegion}
\title{Plot CNV region volcano plot of expression and variation}
\usage{
PlotCNVRegion(object, topn.label = 10, var.limit = 0.001, mean.limit = 0.01)
}
\arguments{
\item{object}{A Seurat object with CNV_region slot}
}
\value{
Returns ggplot2 object
}
\description{
Plot CNV region volcano plot of expression and variation
}
