% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotCVTest}
\alias{PlotCVTest}
\title{Plot CV violin plot of 2 objects divide by reference object}
\usage{
PlotCVTest(object.1, object.2, object.ref, show.exp = F, region = NA)
}
\arguments{
\item{object.1}{A Seurat object with CNV_region slot}

\item{object.2}{A Seurat object with CNV_region slot}

\item{object.ref}{A reference Seurat object with CNV_region slot}

\item{show.exp}{Show expression level of each gene or not}

\item{region}{Specify regions or use all region}
}
\value{
Returns ggplot2 object
}
\description{
Plot CV violin plot of 2 objects divide by reference object
}
