% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotExpCNV}
\alias{PlotExpCNV}
\title{Plot Expression level vs CNV scatter plot}
\usage{
PlotExpCNV(object, gene, region = NULL, plot = T)
}
\arguments{
\item{object}{A Seurat object with CNV_region slot}

\item{gene}{gene name to plot}
}
\value{
Returns ggplot2 object
}
\description{
Plot Expression level vs CNV scatter plot
}
