% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{PlotMST}
\alias{PlotMST}
\title{Plot MST of CNV sub clusters}
\usage{
PlotMST(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns ggplot2 object or development pair and weight
}
\description{
Plot minimal spanning tree of CNV sub clusters
}
