% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{ReadCNR}
\alias{ReadCNR}
\title{Read CNR from .cnr files}
\usage{
ReadCNR(dir, chr.use = paste0("chr", c(seq(1, 22), "X")))
}
\arguments{
\item{chr.use}{You can specify chromosome to use}

\item{CNR}{A dir which contain cnr files}
}
\value{
Returns a CNR object from those files
}
\description{
Read all .cnr files from a dir and output a CNR object
}
