% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{RunCNVPCA}
\alias{RunCNVPCA}
\title{Do PCA to weighted segment matrix}
\usage{
RunCNVPCA(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a CNV matrix or Seurat object
}
\description{
Use PCA to reduce dimension segments with weight
}
