% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{RunCNVUMAP}
\alias{RunCNVUMAP}
\title{Use UMAP to reduce dimension CNV matrix}
\usage{
RunCNVUMAP(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a UMAP reduction or Seurat object
}
\description{
Use UMAP to reduce dimension CNV matrix
}
