% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{SubsetCNR}
\alias{SubsetCNR}
\title{subset CNR object}
\usage{
SubsetCNR(CNR_obj, cells)
}
\arguments{
\item{CNR_obj}{A CNR object}

\item{cells}{Cells you want to keep}
}
\value{
Returns a CNR object after subset
}
\description{
Remove some cells you want to remove
}
